Various changes to support transmission of pictures from the app to the pong central

+++ display_driver/gfx_local_mod/nrf_gfx.h
@@ -243,7 +243,7 @@ void nrf_gfx_screen_fill(nrf_lcd_t const * p_instance, uint32_t color);
  */
 ret_code_t nrf_gfx_bmp565_draw(nrf_lcd_t const * p_instance,
                                nrf_gfx_rect_t const * p_rect,
-                               uint16_t const * img_buf);
+                               uint8_t const * img_buf);
 
 /**
  * @brief Function for drawing an image from a .bmp file.
